再判定ができるように変更した。
git-svn-id: file:///home/moriyoshi/chaperone.e-cell.org/var/opt/subversion/ecell-ide/trunk@2932 fe1fbaf7-ad0a-4b3d-a902-8cb9c273ac6b

+++ ide/EcellLib/EcellLib/Job/JobGroup.cs
@@ -485,6 +485,8 @@ namespace Ecell.Job
                 {
                     string filename = Path.GetFileName(srcname);
                     string dstname = logdir + "/" + filename;
+                    if (!File.Exists(srcname))
+                        continue;
                     File.Copy(srcname, dstname);
                 }
             }
